--- observer/stream/README.txt
@@ -1,1 +1,0 @@
-Not prepared and tested Dockerfile and run_container.sh yet.
